Throw an exception when multipart options is misformatted (#511)
I spend a lot of time searching for an error that was only a misformatted data in the multipart request option, and the server just die without give any information about of error. I think that an exception would be useful for other people that is as absent-minded as me.